net: Add a top level NET_VENDOR_RENESAS
Since both the Synopsys GMACs (STMMAC) and the MoreThanIP 5-port
switch create net devices (ethX), we have a problem with probe
ordering. We want the Synopsys GMACs to get eth0 and eth1, so this
driver has to probe first.

Second, we have a run-time dependency that the Renesas RGMII/GMII
Converters must be probed after the MoreThanIP 5-port switch. This
is because the RGMII/GMII Converters need to hook into PHYs that have
already been probed.

This change does that. It's not nice, I don't like it, but here we are.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>
Signed-off-by: Michel Pollet <michel.pollet@bp.renesas.com>